+++ src/main/java/duke/action/Action.java
@@ -1,4 +1,4 @@
-package duke;
+package duke.action;
 
 public class Action {
     protected final String act;
