+++ src/main/java/duke/command/MarkCommand.java
@@ -1,6 +1,5 @@
 package duke.command;
 
-import duke.Action;
 import duke.Storage;
 import duke.TaskList;
 import duke.Ui;
